// apps/web/src/lib/api/generated/main/model/npc-type.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type NpcType = typeof NpcType[keyof typeof NpcType];


export const NpcType = {
  COMMON: 'COMMON',
  ELITE: 'ELITE',
  ELITE2: 'ELITE2',
  ELITE3: 'ELITE3',
  HERO: 'HERO',
  EVENT_HERO: 'EVENT_HERO',
  TITAN: 'TITAN',
  COLOSSUS: 'COLOSSUS',
  NPC: 'NPC',
} as const;